MAJOR: Add option PersistentTransactions
Configuration client now has an option PersistentTransactions in
ClientParams, and it defaults to true. When true, every change
made inside the transaction is dumped to a transaction file,
otherwise the whole transaction is kept in memory, just in c.parsers
map and dumped to a transaction file only on transaction commit.
This way if performance is needed we can greatly reduce disk I/O.