BUG/MEDIUM: runtime-maps: fix delete entry by its id
when deleting a runtime map entry, it is possible to delete it by its ref or key.
Internally `get map` was used to check for the presence of an entry.
Since it doesn't work with #<ref> syntax, this check is removed.
Now, it is possible to delete an entry by its ref or key.