REORG: bind: enforcing bind `name` as required option mooved from specification to function validation
This change allows reusing bind options in stats socket line.
While `name` property is required for `bind` line, it is optional in `stats socket` line.
Because of that, enforcing `name` property logic is mooved from the specification to the respected
function in bind, so it is possible to use same model in both.