BUG/MEDIUM: fix haproxy version with multiple runtime clients
As haproxy version was a field of the client struct and the version set in a sync.Once function, the version was nil for all runtime clients except the first one. It could also happend that the first call was done with DoNotCheckRuntimeOnInit option and then, it could also happen that even the first one was nil.