BUG/MINOR: Fix serialization of some tunables
Commit ccef2d16 in client-native converted some tunables to Size types,
but this caused an issue with:
 - tune.bufsize
 - tune.pipesize
 - tune.recv_enough

as those variables are stored internally as an int64 instead of a
*int64. This means those variables would be zero-initialized and the
serializeSizeOption() function would write them to the configuration
file with a value of "0". This is rejected by the HAProxy parser which
expected a value > 0.

So this commit modifies serializeSizeOption() to not write zero values,
same as serializeInt64Option() which was used before.