BUG/MEDIUM: Fix bind updates without restart
Some keywords in bind parameters were not using the right type for
serialization (ServerOption instead of BindOption). Since both types
share the same interface, the serialization to disk was working, but
the parser was silently ignoring ServerOptions. So when the user used
the API to create a new Bind and then GET it, the parser would miss
those keywords because they have the wrong type in the cache.

Note that this bug impacts both the deprecated API (like ForceNoSslv3)
and the new API.